chore(CodeBlock): Split out subcomponents
This is purely code organisation.
