+++ src/components/CodeBlock/Lines.tsx
@@ -1,0 +1,40 @@
+import { Box, Stack } from 'braid-design-system';
+import Highlight from 'prism-react-renderer';
+import React from 'react';
+import { useStyles } from 'sku/react-treat';
+
+import { CodeSize } from '../../private/size';
+
+import * as styleRefs from '../CodeBlock.treat';
+
+type Token = Parameters<Highlight['getTokenProps']>[0]['token'];
+
+interface Props {
+  codeSize: CodeSize;
+  getTokenProps: Highlight['getTokenProps'];
+  lines: Token[][];
+}
+
+export const Lines = ({ codeSize, getTokenProps, lines }: Props) => {
+  const styles = useStyles(styleRefs);
+
+  return (
+    <Box padding="medium">
+      <Stack space="small">
+        {lines.map((line, lineIndex) => (
+          <Box
+            className={styles.code[codeSize]}
+            component="pre"
+            key={lineIndex}
+          >
+            {line.map((token, tokenIndex) => {
+              const props = getTokenProps({ token });
+
+              return <Box component="span" {...props} key={tokenIndex} />;
+            })}
+          </Box>
+        ))}
+      </Stack>
+    </Box>
+  );
+};
